Enable overflow tests
