+++ flux-tests/tests/neg/config/test-config00.rs
@@ -1,7 +1,7 @@
 #![feature(register_tool)]
 #![register_tool(flux)]
 #![feature(custom_inner_attributes)]
-#![flux::cfg(log_dir = "./log", dump_constraint = "do it!")] //~ ERROR invalid flux configuration: incorrect type in value for setting `dump_constraint`, expected bool
+#![flux::cfg(check_overflow = "do it!")] //~ ERROR invalid flux configuration: incorrect type in value for setting `check_overflow`, expected bool
 
 #[flux::sig(fn(x: i32, y: i32) -> i32)]
 pub fn test(x: i32, y: i32) -> i32 {
