+++ flux-tests/tests/neg/surface/binop_overflow.rs
@@ -1,8 +1,6 @@
-// ignore-test
-// TODO: Uncomment when we can set flags on a per-test basis. Add the ERROR
-// overflow comment on the lines too.
-#![feature(register_tool)]
+#![feature(register_tool, custom_inner_attributes)]
 #![register_tool(flux)]
+#![flux::cfg(check_overflow = true)]
 
 // Arithmetic BinOps
 // These check for over/underflow
